+++ src/indicators/minimum.rs
@@ -2,8 +2,8 @@ use core::f64::INFINITY;
 use core::fmt;
 use heapless::{Vec, consts::U10};
 
-use crate::errors::*;
-use crate::{Low, Next, Reset};
+use crate::errors::{Error, ErrorKind, Result};
+use crate::{Low, Next, Period, Reset};
 #[cfg(feature = "serde")]
 use serde::{Deserialize, Serialize};
 
@@ -11,7 +11,7 @@ use serde::{Deserialize, Serialize};
 ///
 /// # Parameters
 ///
-/// * _n_ - size of the time frame (integer greater than 0). Default value is 14.
+/// * _period_ - size of the time frame (integer greater than 0). Default value is 14.
 ///
 /// # Example
 ///
@@ -28,35 +28,30 @@ use serde::{Deserialize, Serialize};
 #[cfg_attr(feature = "serde", derive(Serialize, Deserialize))]
 #[derive(Debug, Clone)]
 pub struct Minimum {
-    n: usize,
-    vec: Vec<f64,U10>,
+    period: usize,
     min_index: usize,
     cur_index: usize,
+    deque: Vec<f64,U10>,
 }
 
 impl Minimum {
-    pub fn new(n: u32) -> Result<Self> {
-        let n = n as usize;
-
-        if n <= 0 {
-            return Err(Error::from_kind(ErrorKind::InvalidParameter));
+    pub fn new(period: usize) -> Result<Self> {
+        match period {
+            0 => Err(Error::from_kind(ErrorKind::InvalidParameter)),
+            _ => Ok(Self {
+                period,
+                min_index: 0,
+                cur_index: 0,
+                deque: Vec::new(),
+            }),
         }
-
-        let indicator = Self {
-            n,
-            vec: Vec::new(),
-            min_index: 0,
-            cur_index: 0,
-        };
-
-        Ok(indicator)
     }
 
     fn find_min_index(&self) -> usize {
         let mut min = ::core::f64::INFINITY;
         let mut index: usize = 0;
 
-        for (i, &val) in self.vec.iter().enumerate() {
+        for (i, &val) in self.deque.iter().enumerate() {
             if val < min {
                 min = val;
                 index = i;
@@ -67,25 +62,31 @@ impl Minimum {
     }
 }
 
+impl Period for Minimum {
+    fn period(&self) -> usize {
+        self.period
+    }
+}
+
 impl Next<f64> for Minimum {
     type Output = f64;
 
     fn next(&mut self, input: f64) -> Self::Output {
-        self.vec[self.cur_index] = input;
+        self.deque[self.cur_index] = input;
 
-        if input < self.vec[self.min_index] {
+        if input < self.deque[self.min_index] {
             self.min_index = self.cur_index;
         } else if self.min_index == self.cur_index {
             self.min_index = self.find_min_index();
         }
 
-        self.cur_index = if self.cur_index + 1 < self.n as usize {
+        self.cur_index = if self.cur_index + 1 < self.period {
             self.cur_index + 1
         } else {
             0
         };
 
-        self.vec[self.min_index]
+        self.deque[self.min_index]
     }
 }
 
@@ -99,8 +100,8 @@ impl<T: Low> Next<&T> for Minimum {
 
 impl Reset for Minimum {
     fn reset(&mut self) {
-        for i in 0..self.n {
-            self.vec[i] = INFINITY;
+        for i in 0..self.period {
+            self.deque[i] = INFINITY;
         }
     }
 }
@@ -113,7 +114,7 @@ impl Default for Minimum {
 
 impl fmt::Display for Minimum {
     fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        write!(f, "MIN({})", self.n)
+        write!(f, "MIN({})", self.period)
     }
 }