+++ src/indicators/standard_deviation.rs
@@ -2,8 +2,8 @@ use core::fmt;
 
 use heapless::{Vec, consts::U10};
 use m::Float;
-use crate::errors::*;
-use crate::{Close, Next, Reset};
+use crate::errors::{Error, ErrorKind, Result};
+use crate::{Close, Next, Period, Reset};
 #[cfg(feature = "serde")]
 use serde::{Deserialize, Serialize};
 
@@ -23,7 +23,7 @@ use serde::{Deserialize, Serialize};
 ///
 /// # Parameters
 ///
-/// * _n_ - number of periods (integer greater than 0)
+/// * _period_ - number of periods (integer greater than 0)
 ///
 /// # Example
 ///
@@ -43,29 +43,26 @@ use serde::{Deserialize, Serialize};
 #[cfg_attr(feature = "serde", derive(Serialize, Deserialize))]
 #[derive(Debug, Clone)]
 pub struct StandardDeviation {
-    n: u32,
+    period: usize,
     index: usize,
-    count: u32,
+    count: usize,
     m: f64,
     m2: f64,
-    vec: Vec<f64, U10>,
+    deque: Vec<f64, U10>,
 }
 
 impl StandardDeviation {
-    pub fn new(n: u32) -> Result<Self> {
-        match n {
+    pub fn new(period: usize) -> Result<Self> {
+        match period {
             0 => Err(Error::from_kind(ErrorKind::InvalidParameter)),
-            _ => {
-                let std = StandardDeviation {
-                    n,
-                    index: 0,
-                    count: 0,
-                    m: 0.0,
-                    m2: 0.0,
-                    vec: Vec::new(),
-                };
-                Ok(std)
-            }
+            _ => Ok(Self {
+                period,
+                index: 0,
+                count: 0,
+                m: 0.0,
+                m2: 0.0,
+                deque: Vec::new(),
+            }),
         }
     }
 
@@ -74,20 +71,26 @@ impl StandardDeviation {
     }
 }
 
+impl Period for StandardDeviation {
+    fn period(&self) -> usize {
+        self.period
+    }
+}
+
 impl Next<f64> for StandardDeviation {
     type Output = f64;
 
     fn next(&mut self, input: f64) -> Self::Output {
-        let old_val = self.vec[self.index];
-        self.vec[self.index] = input;
+        let old_val = self.deque[self.index];
+        self.deque[self.index] = input;
 
-        self.index = if self.index + 1 < self.n as usize {
+        self.index = if self.index + 1 < self.period {
             self.index + 1
         } else {
             0
         };
 
-        if self.count < self.n {
+        if self.count < self.period {
             self.count += 1;
             let delta = input - self.m;
             self.m += delta / self.count as f64;
@@ -96,9 +99,12 @@ impl Next<f64> for StandardDeviation {
         } else {
             let delta = input - old_val;
             let old_m = self.m;
-            self.m += delta / self.n as f64;
+            self.m += delta / self.period as f64;
             let delta2 = input - self.m + old_val - old_m;
             self.m2 += delta * delta2;
+        }
+        if self.m2 < 0.0 {
+            self.m2 = 0.0;
         }
 
         (self.m2 / self.count as f64).sqrt()
@@ -119,8 +125,8 @@ impl Reset for StandardDeviation {
         self.count = 0;
         self.m = 0.0;
         self.m2 = 0.0;
-        for i in 0..(self.n as usize) {
-            self.vec[i] = 0.0;
+        for i in 0..self.period {
+            self.deque[i] = 0.0;
         }
     }
 }
@@ -133,7 +139,7 @@ impl Default for StandardDeviation {
 
 impl fmt::Display for StandardDeviation {
     fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        write!(f, "SD({})", self.n)
+        write!(f, "SD({})", self.period)
     }
 }
 
@@ -159,6 +165,18 @@ mod tests {
         assert_eq!(round(sd.next(20.0)), 7.071);
         assert_eq!(round(sd.next(10.0)), 7.071);
         assert_eq!(round(sd.next(100.0)), 35.355);
+    }
+
+    #[test]
+    fn test_next_floating_point_error() {
+        let mut sd = StandardDeviation::new(6).unwrap();
+        assert_eq!(sd.next(1.872), 0.0);
+        assert_eq!(round(sd.next(1.0)), 0.436);
+        assert_eq!(round(sd.next(1.0)), 0.411);
+        assert_eq!(round(sd.next(1.0)), 0.378);
+        assert_eq!(round(sd.next(1.0)), 0.349);
+        assert_eq!(round(sd.next(1.0)), 0.325);
+        assert_eq!(round(sd.next(1.0)), 0.0);
     }
 
     #[test]